marketing_ai/pyproject.toml
[project]
name = "marketing_ai"
version = "0.1.0"
description = "marketing-ai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0",
    "fpdf2"
]

[project.scripts]
marketing_ai = "marketing_ai.main:run"
run_crew = "marketing_ai.main:run"
train = "marketing_ai.main:train"
replay = "marketing_ai.main:replay"
test = "marketing_ai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
